Move move_checker to separate file
